+++ CHANGELOG.md
@@ -21,7 +21,8 @@ This version add the orders module
 
 ### Added
 - Added all services operations in `src/api/v1/apps/orders/services.py`
-- Added enpoints for CRUD orders and list templates in `src/api/v1/apps/orders/router.py`
+- Added endpoints for CRUD orders and list templates in `src/api/v1/apps/orders/router.py`
+- Added endpoint for products list in `src/api/v1/apps/orders/router.py`
 
 ### Changed
 - Added current_user at context in patch.products endpoint in `src/api/v1/apps/orders/router.py`